Refactored API to use new database
