+++ src/spec/apis/funds_api.py
@@ -52,7 +52,7 @@ def _get_fastapi_sessionmaker() -> FastAPISessionMaker:
     Returns:
         FastAPISessionMaker: FastAPI session maker
     """
-    database_uri = os.environ["SQLALCHEMY_DATABASE_URL"]
+    database_uri = os.environ["DATABASE_URL"]
     return FastAPISessionMaker(database_uri)
