+++ src/tests/data/backend-funds.sql
@@ -1,0 +1,6 @@
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("03568d76-93f1-3a2d-9ed5-b95516546548", "-",""))), 123, 'PASSIVETEST01', 'Passive test fund 1 - fi', 'Passive test fund 1 - en');
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("dc856547-449b-306e-80c9-05ef8a002a3a", "-",""))), 234, 'ACTIVETEST01', 'Active test fund 1 - fi', 'Active test fund 1 - en');
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("3073d4fe-78cc-36ec-a7e0-9c24944030b0", "-",""))), 345, 'BALANCEDTEST01', 'Balanced test fund 1 - fi', 'Balanced test fund 1 - en');
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("8cd9b437-e524-3926-a3f9-969923cf51bd", "-",""))), 456, 'FIXEDTEST01', 'Fixed test fund 1 - fi', 'Fixed test fund 1 - en');
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("79f8da68-6bdc-3a24-acde-327aa14e2546", "-",""))), 678, 'DIMETEST01', 'Dimensional test fund 1 - fi', 'Dimensional test fund 1 - en');
+INSERT INTO fund (id, fund_id, security_id, security_name_fi, security_name_sv) VALUES ((UNHEX(REPLACE("098a999d-65ae-3746-900b-b0b33a5d7d9c", "-",""))), 789, 'SPILTAN TEST', 'Spiltan test fund 1 - fi', 'Spiltan test fund 1 - en');
