+++ src/tests/fixtures/client.py
@@ -3,6 +3,7 @@ import os
 
 from starlette.testclient import TestClient
 from ...app.main import app
+from testcontainers.mysql import MySqlContainer
 
 data_folder = os.path.join(os.path.dirname(__file__), '..', 'data')
 funds_json = os.path.join(data_folder, 'funds.json')
@@ -11,7 +12,7 @@ fund_values_basic_csv = os.path.join(data_folder, 'fund-values-basic.csv')
 
 
 @pytest.fixture()
-def client() -> TestClient:
+def client(backend_mysql: MySqlContainer) -> TestClient:
     """Fixture for test REST client
 
     Returns:
