+++ src/tests/test_funds_api.py
@@ -1,4 +1,5 @@
 import logging
+import time
 
 from typing import List
 from starlette.testclient import TestClient
@@ -7,7 +8,12 @@ from .auth.auth import BearerAuth
 
 from .fixtures.client import *  # noqa
 from .fixtures.users import *  # noqa
-from .fixtures.mysql import *  # noqa
+from .fixtures.salkku_mysql import *  # noqa
+from .fixtures.backend_mysql import *  # noqa
+from .fixtures.sync import *  # noqa
+from .fixtures.kafka import *  # noqa
+from .fixtures.kafka_connect import *  # noqa
+from .fixtures.zookeeper import *  # noqa
 
 logger = logging.getLogger(__name__)
 
@@ -26,7 +32,7 @@ invalid_uuids = ["potato", "`?%!", "äö", "Правда"]
 class TestFunds:
     """Tests for funds endpoints"""
 
-    def test_find_fund(self, client: TestClient, user_1_auth: BearerAuth, mysql: MySqlContainer):
+    def test_find_fund(self, client: TestClient, user_1_auth: BearerAuth):
         fund_id = fund_ids["passivetest01"]
         response = client.get(f"/v1/funds/{fund_id}", auth=user_1_auth)
         assert response.status_code == 200
@@ -55,13 +61,13 @@ class TestFunds:
         assert -0.32 == fund["profitProjection"]
         assert "2021-10-12" == fund["profitProjectionDate"]
 
-    def test_find_fund_invalid_id(self, client: TestClient, user_1_auth: BearerAuth, mysql: MySqlContainer):
+    def test_find_fund_invalid_id(self, client: TestClient, user_1_auth: BearerAuth):
         for invalid_uuid in invalid_uuids:
             url = f"/v1/funds/{invalid_uuid}"
             response = client.get(url, auth=user_1_auth)
             assert response.status_code == 400
 
-    def test_list_funds(self, client: TestClient, user_1_auth: BearerAuth, mysql: MySqlContainer):
+    def test_list_funds(self, client: TestClient, user_1_auth: BearerAuth):
         response = client.get("/v1/funds", auth=user_1_auth)
         assert response.status_code == 200
 
@@ -71,7 +77,7 @@ class TestFunds:
         for fund_id in fund_ids.values():
             assert fund_id in response_ids
 
-    def test_list_funds_limits(self, client: TestClient, user_1_auth: BearerAuth, mysql: MySqlContainer):
+    def test_list_funds_limits(self, client: TestClient, user_1_auth: BearerAuth):
 
         self.assert_list(
             client=client,
@@ -109,7 +115,7 @@ class TestFunds:
             ]
         )
 
-    def test_find_fund_values(self, client: TestClient, user_1_auth: BearerAuth, mysql: MySqlContainer):
+    def test_find_fund_values(self, client: TestClient, user_1_auth: BearerAuth):
         fund_id = fund_ids["passivetest01"]
         start_date = "2020-01-01"
         end_date = "2020-01-05"
